+++ CHANGES.md
@@ -396,3 +396,4 @@
 ### Pre-release (work-in-progress)
 - Quick Scan: add back Starkey Inspire
 - Phonak: workaround for Roger Upgrader checker in case relative path is used on its webpage
+- ReSound/Beltone/Interton/Danavox: update archived versions