+++ CHANGES.md
@@ -397,3 +397,4 @@
 - Quick Scan: add back Starkey Inspire
 - Phonak: workaround for Roger Upgrader checker in case relative path is used on its webpage
 - ReSound/Beltone/Interton/Danavox: update archived versions
+- Starkey: add archived versions