+++ CHANGES.md
@@ -398,3 +398,5 @@
 - Phonak: workaround for Roger Upgrader checker in case relative path is used on its webpage
 - ReSound/Beltone/Interton/Danavox: update archived versions
 - Starkey: add archived versions
+- Widex: fix market fallback bug
+- Widex: add archived versions